Created custom OGR osmconf.ini file and adjusted PBF driver option to work correctly
Generated a custom osmconf.ini file, which causes the OGR driver to
expose the fields required for Mapaction format output. However as yet
the code does not reference this custom file "in situ" and it has to be
copied to the gdal directory (C:\OSGeo4W64\share\gdal) to work.
Additionally the code then only copied features to the point shapefiles.
This has been temporarily fixed by turning off the
OGR_INTERLEAVED_READING option but that may have other consequences so
it is still considered an experimental fix!